+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,17 @@
 # effect-solutions
 
+## 0.5.1
+
+### Patch Changes
+
+- [`10a45a2`](https://github.com/kitlangton/effect-solutions/commit/10a45a25ab6e1b386e1867ab2d729fcd6ad367a9) Thanks [@kitlangton](https://github.com/kitlangton)! - Fix CLI argv parsing for Effect v4 runWith
+
+## 0.5.0
+
+### Minor Changes
+
+- [`98ca943`](https://github.com/kitlangton/effect-solutions/commit/98ca943117cdef4110547cbd353c6927f9df586e) Thanks [@kitlangton](https://github.com/kitlangton)! - remove setup command update docs for v4
+
 ## 0.4.19
 
 ## 0.4.18