+++ packages/website/CHANGELOG.md
@@ -1,5 +1,9 @@
 # @effect-best-practices/website
 
+## 0.5.1
+
+## 0.5.0
+
 ## 0.4.19
 
 ### Patch Changes